Moved each sub component to a different file
Moved some constants to the presenter file
