+++ src/components/colorPicker/ColorPickerDialogHeader.tsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import _ from 'lodash';
+import {View, Button} from 'react-native-ui-lib';
+import {StyleSheet} from 'react-native';
+import Assets from '../../assets';
+import {Colors} from '../../style';
+import {ColorPickerDialogProps} from './ColorPickerDialog';
+import {BORDER_RADIUS} from './ColorPickerPresenter';
+type HeaderProps = Pick<ColorPickerDialogProps, 'doneButtonColor' | 'accessibilityLabels' | 'testID'> & {
+  valid: boolean;
+  onDismiss: () => void;
+  onDonePressed: () => void;
+};
+
+const Header = (props: HeaderProps) => {
+  const {onDismiss, accessibilityLabels, testID, doneButtonColor, valid, onDonePressed} = props;
+
+  return (
+    <View row spread bg-white paddingH-20 style={styles.header}>
+      <Button
+        link
+        iconSource={Assets.icons.x}
+        iconStyle={{tintColor: Colors.$iconDefault}}
+        onPress={onDismiss}
+        accessibilityLabel={_.get(accessibilityLabels, 'dismissButton')}
+        testID={`${testID}.dialog.cancel`}
+      />
+      <Button
+        color={doneButtonColor}
+        disabled={!valid}
+        link
+        iconSource={Assets.icons.check}
+        onPress={onDonePressed}
+        accessibilityLabel={_.get(accessibilityLabels, 'doneButton')}
+        testID={`${testID}.dialog.done`}
+      />
+    </View>
+  );
+};
+
+export default Header;
+
+const styles = StyleSheet.create({
+  header: {
+    height: 56,
+    borderTopLeftRadius: BORDER_RADIUS,
+    borderTopRightRadius: BORDER_RADIUS,
+    backgroundColor: Colors.$backgroundDefault
+  }
+});
